feat: Implement worker threads for upload processing
- Offload flashcard generation to a separate worker thread
- Prevents blocking of main thread during uploads, improving responsiveness
- Addresses bug causing server slowdown during large uploads

Refs:
https://nodejs.org/api/worker_threads.html
https://stackoverflow.com/questions/61095741/how-do-i-avoid-blocking-an-express-rest-service
